simplecrm/requirements.txt
Django==2.1.3
django-extensions==2.1.3
djangorestframework==3.9.0
gunicorn==19.9.0
pkg-resources==0.0.0
pytz==2018.7
six==1.11.0
whitenoise==4.1
